+++ src/router.jsx
@@ -6,11 +6,11 @@ import CircularProgress from "@mui/material/CircularProgress"
 const Login = lazy(() => import("@/modules/Login"))
 const Register = lazy(() => import("@/modules/Register"))
 const NotFound = lazy(() => import("@/modules/NotFound"))
-const HomePage = lazy(() => import("@/modules/pages/HomePage"))
+const MainPage = lazy(() => import("@/modules/pages/MainPage"))
 const TestPage = lazy(() => import("@/modules/pages/TestPage"))
 
 const routes = [
-	{ path: "/", element: <HomePage />, index: true },
+	{ path: "/", element: <MainPage />, index: true },
 	{ path: "/login", element: <Login /> },
 	{ path: "/test", element: <TestPage /> },
 	{ path: "/register", element: <Register /> },
